api-server/covenant-service/requirements.txt
fastapi
uvicorn
sqlalchemy==1.4.23
asyncpg
pytz
python-dotenv
psycopg2-binary
email-validator
passlib[bcrypt]==1.7.4
shortuuid==1.0.13
python-jose[cryptography]
redis[asyncio]>=4.5.5


